+++ src/main/java/com/example/commonproject/payment/controller/SettlementController.java
@@ -1,0 +1,13 @@
+package com.example.commonproject.payment.controller;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Controller;
+
+/**
+ * 정산
+ */
+@Slf4j
+@Controller
+public class SettlementController {
+
+}
