+++ src/main/java/com/example/commonproject/payment/service/OrderService.java
@@ -1,0 +1,7 @@
+package com.example.commonproject.payment.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class OrderService {
+}
