+++ src/main/java/com/example/commonproject/payment/util/OrderUtil.java
@@ -1,0 +1,4 @@
+package com.example.commonproject.payment.util;
+
+public class OrderUtil {
+}
