+++ src/main/java/com/example/commonproject/payment/util/PaymentUtil.java
@@ -1,0 +1,4 @@
+package com.example.commonproject.payment.util;
+
+public class PaymentUtil {
+}
